Add possibility to show list of abandoned colonies
since hours
