+++ src/main/java/com/minecolonies/coremod/commands/CitizensCommand.java
@@ -1,6 +1,10 @@
 package com.minecolonies.coremod.commands;
 
 import com.google.common.collect.ImmutableMap;
+import com.minecolonies.coremod.commands.citizencommands.CitizenInfoCommand;
+import com.minecolonies.coremod.commands.citizencommands.KillCitizenCommand;
+import com.minecolonies.coremod.commands.citizencommands.ListCitizensCommand;
+import com.minecolonies.coremod.commands.citizencommands.RespawnCitizenCommand;
 import org.jetbrains.annotations.NotNull;
 
 import java.util.Map;
