+++ src/main/java/com/minecolonies/coremod/commands/ColonyCommand.java
@@ -1,6 +1,7 @@
 package com.minecolonies.coremod.commands;
 
 import com.google.common.collect.ImmutableMap;
+import com.minecolonies.coremod.commands.colonycommands.*;
 import org.jetbrains.annotations.NotNull;
 
 import java.util.Map;
@@ -22,6 +23,8 @@ public class ColonyCommand extends AbstractSplitCommand
         .put(RefreshColonyCommand.DESC, new RefreshColonyCommand(MinecoloniesCommand.DESC, ColonyCommand.DESC, RefreshColonyCommand.DESC))
         .put(ChangeColonyOwnerCommand.DESC, new ChangeColonyOwnerCommand(MinecoloniesCommand.DESC, ColonyCommand.DESC, ChangeColonyOwnerCommand.DESC))
         .put(ColonyTeleportCommand.DESC, new ColonyTeleportCommand(MinecoloniesCommand.DESC, ColonyCommand.DESC, ColonyTeleportCommand.DESC))
+        .put(FindAbandonedColoniesCommand.DESC, new FindAbandonedColoniesCommand(MinecoloniesCommand.DESC, ColonyCommand.DESC, FindAbandonedColoniesCommand.DESC))
+
               .build();
 
     /**
