+++ src/main/java/com/minecolonies/coremod/commands/MinecoloniesCommand.java
@@ -1,6 +1,9 @@
 package com.minecolonies.coremod.commands;
 
 import com.google.common.collect.ImmutableMap;
+import com.minecolonies.coremod.commands.colonycommands.HomeTeleportCommand;
+import com.minecolonies.coremod.commands.generalcommands.BackupCommand;
+import com.minecolonies.coremod.commands.generalcommands.RandomTeleportCommand;
 import com.minecolonies.coremod.configuration.Configurations;
 import net.minecraft.entity.player.EntityPlayer;
 import org.jetbrains.annotations.NotNull;
