+++ src/main/java/com/minecolonies/coremod/commands/citizencommands/ListCitizensCommand.java
@@ -1,9 +1,10 @@
-package com.minecolonies.coremod.commands;
+package com.minecolonies.coremod.commands.citizencommands;
 
 import com.minecolonies.coremod.colony.CitizenData;
 import com.minecolonies.coremod.colony.Colony;
 import com.minecolonies.coremod.colony.ColonyManager;
 import com.minecolonies.coremod.colony.IColony;
+import com.minecolonies.coremod.commands.AbstractSingleCommand;
 import net.minecraft.command.CommandException;
 import net.minecraft.command.ICommandSender;
 import net.minecraft.entity.player.EntityPlayer;
