+++ src/main/java/com/minecolonies/coremod/commands/colonycommands/package-info.java
@@ -1,0 +1,4 @@
+/**
+ * Contains all commands related to colonies.
+ */
+package com.minecolonies.coremod.commands.colonycommands;
